requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
llama-cpp-python
langchain
faiss-cpu
sentence-transformers
chromadb
llama-index